cros-kernel2: Add support for wireless38 USE flag
The new wireless38 use flag is used to specify whether a kernel
should build with the old 3.8 wifi driver/stack or with the newer
3.10 one.

It is expected that overlays that want to stay back with 3.8 drivers
set the USE flag in their make.conf, and that others will continue to
move forward.

Since the wireless34 flag conflicts with this, make sure that an
error is reported if this is attempted.

BUG=chromium:306710
TEST=Emerge the following:
 - 'emerge-falco chromeos-kernel-3_10' and ensure WIFIVERSION=-3.4 is
    applied to the build (since this appears in falco's make.conf).
 - 'USE=wireless38 emerge-falco chromeos-kernel-3_10' and ensure
    an error is reported "Wireless 3.8 AND 3.4 both set"
 - 'USE="wireless38 -wireless34" emerge-falco chromeos-kernel-3_10' and
    ensure WIFIVERSION=-3.8 is applied to the build.

Change-Id: I11854a6987eaa35feff25f92f382c9469101d350
Reviewed-on: https://chromium-review.googlesource.com/177595
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>