coreboot: Use the fwserial.[board] file to enable the serial console.
Instead of adding a hardcoded list of options in the board specific ebuild, we
can append a file which has the options in it which is stored alongside the
coreboot configs. This helps pull things out of the board specific ebuilds and
keeps the configs in one place.

BUG=None
TEST=Built and booted on nyan. Built for all other boards except baskingridge
which is currently broken.
BRANCH=None

Change-Id: I7243126dc3db9e762fa02c3b779a126a6aeb6f47
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/178220
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>