Add Pit Xorg config to the build for Daisy
There are Daisy's out there with two different touchpads.  One is the
original Cypress pad, the other is the new Atmel one.  The Atmel modules
are exactly the same as Pit's.  For the sake of simplicity, this patch
simply has the xorg-conf ebuild copy in the pit xorg-conf as well when
it builds for Lucas instead of having to duplicate the config for both
devices.  Any bug fix for the Atmel pad on Pit should effect Lucas
and vice versa, and this makes it easier to do that

BUG=chromium:323958
TEST=Gmerged on Lucas with Atmel and everything got copied correctly and worked

Change-Id: I677849e925e7ef085903b8161c5617c6f91ae2a2
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/178203
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>