Create profilers __init__.py in autotest rather than autotest-deps
This CL moves the creation of client/profilers/__init__.py from
autotest-deps.ebuild to autotest.ebuild, so that it may be included in
the client autotest tarball (which will be created in autotest.ebuild in
a future CL). In order to avoid a file collision, this requires a small
hack in the autotest-deps ebuild.

BUG=chromium:324963
TEST=full `build_packages` while working on autotest and autotest-deps
works. Incremental build works.

Change-Id: I6410996fcdb26bf672a0262913b510e464109b38
Reviewed-on: https://chromium-review.googlesource.com/178543
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>