Re-land "Make chromeos-fonts and chromeos-assets dependencies reflect reality"
This reverts commit 935926455aecb0dd5acb3857b1c51824eeeb973e.

chromeos-fonts is needed by chromeos-assets and chromeos-chrome, while
chromeos-init directly requires some tools and things that are installed by
chromeos-assets.

The chromeos-assets package also incorrectly DEPENDed on some font packages
and xcursorgen. These should be (and already are) hard-host-depends.

The chromeos-fonts package ALSO RDEPENDs on chromeos-assets when built for
the target, with no explanation. This seems to be backwards.

BUG=chromium:321784
TEST=build_packages for lumpy, ensure that chromeos-assets and chromeos-fonts are still installed

Change-Id: I41cc33438efb64bf1c18e0296435a44a944d3b2a
Reviewed-on: https://chromium-review.googlesource.com/178701
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>