perf: fix post unwind support and perf trace
Add explicit dependencies for "sys-process/audit" which is used for
"perf trace" command and "sys-libs/libunwind" which is mandatory to get
post unwind support (to have the user call stack for syscalls).

Fix the architecture for the host, ARCH needs to be x86_64 to get the
userspace tool properly compiled (eg the right libunwind flags).

BUG=chromium:327526
TEST=sudo emerge perf-next ; emerge-link perf-next
"CHK libunwind" and "CHK libaudit" are no longer failing during the
build.

Change-Id: Ie00a7e88395b4314032626d42ac6279feac06fcb
Reviewed-on: https://chromium-review.googlesource.com/179606
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>