Add udev rule -- chown input devices' wakeup sysfs
This adds a new udev rule that chown's any new input device's
power/wakeup sysfs entries as they come up to be owned by the user
"power."  For powerd to be able to do it's job it needs to be the owner
of these files, but the way it was done before (in powerd.conf) had a
little race condition that could cause problems every now and then.
This way devices can come and go as they please and still maintain the
correct permissions when they are around.

BUG=chrome-os-partner:24299
TEST=Manual testing on Lucas
CQ-DEPEND=I7bb1a1336f2de7bc87b9f52ae4065ea96eabe933

Change-Id: Ic6b2687a5bcd499e0baefc935117bb928aaaa1ac
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/179847
Reviewed-by: Daniel Erat <derat@chromium.org>