migrate to new qemu package
Upstream has merged all the qemu packages into one.  Update to that
new version.

When trying to move from qemu-kvm to qemu, we run into a circular
upgrade issue: the package is used in the sdk by both cros-workon
and non-cros-workon packages.  In order to run the Uprev phase, we
need an sdk, but we can't initialize the sdk because the qemu pkgs
block each other.  For now add a dummy qemu-kvm package that just
pulls in the new qemu.  We'll drop it in a follow up.

BUG=chromium:220911
TEST=`cbuildbot chromiumos-sdk` works
TEST=`emerge-daisy chromeos-fonts` works (uses qemu)
TEST=ran unittests for platform2 for daisy and it passed (uses qemu)
CQ-DEPEND=CL:180922

Change-Id: I5b5dd94afb0fb06861b527288cd6fbabc22ed88d
Reviewed-on: https://chromium-review.googlesource.com/211090
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>