Set CC_host and CXX_host to clang and clang++ in clang builds.
CC_host and CXX_host are supposed to override make_global_settings,
so these need to be set.  (Due to a bug in the make generator, this
doesn't currently happen, so this change is a prerequiste for fixing
the make generator.  It's also a prerequiste for switching the
chromiumos asan bot to ninja.)

BUG=chromium:326411
TEST=Turn on +ninja in IUSE, no compile error on x86-generic-asan and
amd64-generic-asan.

Change-Id: Iffe7c0fa1267bbbd54b66ef43773adbd5a61a1e6
Reviewed-on: https://chromium-review.googlesource.com/180522
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Tested-by: Nico Weber <thakis@chromium.org>