Give powerd ownership to all wakeup sysfs entries
Previously the udev rule only matched devices with power/wakeup sysfs
entries that came up initially in the "enabled" state.  It turns out
that Atmel touchpads default to disabled, so they were getting skipped
which resulted in powerd being unable to enable and use them as wakeup
sources.

This change alters the rule to match any input device that has a
power/wakeup sysfs entry regardless of it's intial state.

BUG=chromium:24604
TEST=manual testing on a Link (with Atmel touchpad).  I forced a suspend
and the touchpad was able to wake it up.  Additionally, all of the
power/wakeup files were given the correct ownerships.

Change-Id: I5a7e3e422c97a6e8627f4c2bc8ed6f2d9a916668
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/180563
Reviewed-by: Daniel Erat <derat@chromium.org>