Revert "Add egl, gbm, and shared-glapi USE flags to media-libs/mesa so that"
This reverts commit 8c5117605df9506df07b791bace64d9041704667.

We don't really want to be building mesa with EGL support on all x86 boards.
Right now we only want to enable it on {x86,amd64}-generic to allow
experimentation (see [0], although that patch does not reference a BUG,
so the justification discussion is lost to history).

[0] https://gerrit.chromium.org/gerrit/51379

BUG=chromium:329672
TEST=emerge-x86-alex mesa
  => Builds without USE="egl"
TEST=emerge-x86-generic mesa
  => Still builds with USE="egl"

Change-Id: I48c26d52148556fb2e53612debd1af2d5707cd6a
Reviewed-on: https://chromium-review.googlesource.com/180780
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>