app-benchmarks/glmark2: Add an ebuild that builds from upstream tarball
glmark2 is an OpenGL (ES) 2.0 benchmark.  It makes some sense for it to
live in app-benchmarks.

A previous glmark2 ebuild built from a Chromium OS git repository
that was forked from upstream release 2012.01.  It had some build fixes,
and the addition of a DRM canvas. Both the build fixes and the DRM canvas
are now upstream, so for now let's just build from the upstream tarballs.

This second CL adds the new ebuild.  In the future, small patches can be
maintained as ebuild patches, while waiting for them to be included in a
new upstream release.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:330529
TEST=for b in arm-generic link daisy x86-generic x86-alex nyan; do
       emerge-${b} glmark2 ;
     done
   => Should successfully build app-benchmarks/glmark2
CQ-DEPEND=CL:181266

Change-Id: Ife9e2abb86eff149481789fb0f5a3dfd2f7466fc
Reviewed-on: https://chromium-review.googlesource.com/181107
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>