media-libs/glmark2: Remove ebuild that builds from local repo
The Chromium OS glmark2 repository is based off of upstream release
2012.01, with some build fixes, and the addition of a DRM canvas.
Both the build fixes and the DRM canvas are now upstream, so, let's just
build from the upstream binaries instead of carrying a Chromium OS fork.

This first CL removes the old ebuild.  A followup CL will add a new ebuild
to app-benchmarks/glmark2 that builds from an upstream tarball.

In the future, small patches can be maintained as ebuild patches, while
waiting for them to be included in a new upstream release.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:330529
TEST=cbuildbot link-full daisy-full daisy-paladin

Change-Id: I67c10cb705a1493ac58ed8e47df8a2b3ac957e03
Reviewed-on: https://chromium-review.googlesource.com/181106
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>