app-benchmarks/glmark2: Fix --fullscreen
If the Atom "_NET_WM_STATE_FULLSCREEN" is not defined in the X server, it
could mean that the active Window Manager is not EWMH compliant, or that
there is no window manager at all.  In this case, the XChangeProperty
will fail and the program will terminate with the message:

X Error of failed request:  BadAtom (invalid Atom parameter)
  Major opcode of failed request:  18 (X_ChangeProperty)
  Atom id in failed request:  0x0
  Serial number of failed request:  28
  Current serial number in output stream:  30

Instead, if the Atom doesn't exist, don't call XChangeProperty, just log
a message and keep going.  The window will still be fullscreen, it just
may still have decorations.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:330529
TEST=On link:
  echo 0 > /sys/kernel/debug/dri/0/i915_fbc
  glmark2 --fullscreen
  => glmark2 window takes up the whole screen
TEST=On daisy:
 glmark2-es2 --fullscreen --visual-config 'a=0:buf=0'

Change-Id: Ia3793701b1787869e38b612c2a575c8c5efd31d9
Reviewed-on: https://chromium-review.googlesource.com/183270
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>