Remove git/subversion from the provided packages list.
Moblab requires git and subversion to be installed on the device but this
is currently blocked as it is in the provided packages list. This change
removes them from this blacklist.

BUG=chromium:335814
TEST=Try-bot with stumpy-release and hwtest. Logged into a machine with
this image installed and verified git/svn were not avaliable. Then a local
stumpy_moblab build to ensure that on that image they are avaliable.

Change-Id: Ife5613659946a54fca0d534685b03b1164c1fa19
Reviewed-on: https://chromium-review.googlesource.com/183287
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>