app-benchmarks/glmark2: Update to latest version
It turns out upstream stopped publishing tar balls for glmark2 in 2012,
but there has been a years worth of patches added to the bzr repository
since then.

I've taken a snapshot of bzr release 285, and uploaded it to Chromium
localmirror, and updated this ebuild to fetch and build this latest
version.

Note: the upgrade comes with some ebuild fixups:
 * Remove libpng sed command - upstream fixed its png dependency
 * Update 'flavor' selection - configure options changed
 * Require at least one client API
 * Add virtual/jpeg dependency (libjpeg)

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:330529
TEST=for b in daisy link peach_pit x86-generic:
  emerge-${b} glmark2
TEST=On link:
  DISPLAY=:0 glmark2
TEST=On daisy:
  DISPLAY=:0 glmark2-es2

Change-Id: If3ce1e0e2da286c001689a1727680a09879ddf82
Reviewed-on: https://chromium-review.googlesource.com/183234
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>