chromeos-dev: Use upstream x11-misc/xtrace not cros specific x11-apps/xtrace
Now that we've uprev'ed the portage-stable xtrace to 1.3.0, we can pull it
into dev images.

While we are at it, we can get rid of the no-longer-needed x11-apps/xtrace
ebuild in the chromiumos-overlay.
However, in order to avoid file collisions, we need to make the old ebuild
a RDEPEND blocker to get portage to unmerge it before merging the new one.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:330990,chromium:336702
TEST=emerge xtrace
  => builds x11-misc/xtrace
TEST=On link:
  # x11trace -d :0 -D :9 glxinfo | grep DRI2
TEST=On daisy:
  # x11trace -d :0 -D :9 xdriinfo | grep DRI2
=> DRI2 messages are interpreted fully (No UNKNOWN opcodes)

Change-Id: I42cbf669957af8b6271b6a70f88d6fe40130c7c4
Reviewed-on: https://chromium-review.googlesource.com/183297
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>