Add ebuild file for DPTF (Dynamic Platform & Thermal Framework)
This is the initial submission of the ebuild file that builds and
installs the daemon component of DPTF and its related policy shared
libraries.

BUG=none
TEST=manual:

1) run "emerge-rambi <package name>" to ensure that code compiles and
installs correctly.

2) Also run "build_packages.sh" and "build_image.sh" to create the
entire Chromium OS image, and install it on a Rambi board, and check
that DPTF files are installed correctly, and the daemon (esif_ufd) is
started.

Change-Id: Ibdce17ee9df6ffc30c5eda12fc21ca8ea1d1aeae
Signed-off-by: Brian Bian <brian.bian@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/183575
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>