[chromiumos] Make chromeos-chrome depend on sys-apps/pciutils.
The transition of IDs from pciutils/usbutils to hwids means that we've
dropped a dependency on pciutils.  It turns out that chrome depends on
pciutils and that this dependency was always being satisified
transitively, which no longer happens.  Therefore, we need to explicitly
add it.

TEST=ChromeSDK passes trybots
BUG=chromium:333921

Change-Id: I1379db2d60f20e879dfe9c3c01783d1ca1e0575b
Reviewed-on: https://chromium-review.googlesource.com/184251
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>