depthcharge: Communicate mock tpm requirement through defconfig
When USE=mocktpm is set, the depthcharge ebuild sets an environment
variable, which causes the depthcarge Makefile compile vboot_reference
with alternative compilation options.

It is easier to communicate the need to enable mock TPM support
through a depthcharge board deconfig file. This is also more
convenient, since adding more use flags of the kind will not require
modifying the depthcharge Makefile.

BUG=chromium:338061
TEST=manual

   . modified depthcharge Makefile to ignore the environment setting.
     Ran the following commands:

   $ emerge-rambi depthcharge
   $ USE=mocktpm emerge-rambi depthcharge

    and verified that in the second case vboot_reference is indeed
    compiling the appropriate modules.

Change-Id: I9cf2ae4d8bc071d9b9e8fdfa270523921e1476ce
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/184441
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>