app-benchmarks/glmark2: Add --data-path command line option
glmark2 needs access to models, shaders and textures.  These are stored
in the glmark data path.

Currently glmark2 has a --data-path build-time option to specify where
these files are installed at build time.  This option is passed in at
compile time via the GLMARK_DATA_PATH macro.

In some situations, however, it might become necessary to specify the
location of the data files at *runtime*.  This patch adds that option.
If not specified, the compile-time GLMARK_DATA_PATH value is used.

In particular, this feature is used when using glmark2 as a Chrome OS
autotest dependency.  glmark2 is built and installed in the default
locations in a development host chroot (ie ${SYSROOT}/usr/bin/glmark2 and
${SYSROOT}/usr/share/glmark2).  The autotest deps system then rsyncs the
executable and data into a special sandbox on the DUT.  Thus, the data
files will not actually get installed at /usr/share/glmark2, and instead
we use this option to specify their location within the autotest sandbox.

While we are here, also update KEYWORDS="*" to suppress presubmit check.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:330529
TEST=On device:
  gmerge glmark2
  mkdir /tmp/glmark2
  cp -R /usr/share/glmark2 /tmp/glmark2/data
  glmark2 --data-path=/tmp/glmark2/data

Change-Id: I84981ec68c07fafd3af728bd447d4515bd214650
Reviewed-on: https://chromium-review.googlesource.com/184282
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>