xorg-server patch to fix mirror mode unplug/re-plug
The function RRCrtcSet call checks to see if the config being set is
already configured, but, doesn't check that the selected outputs are
connected to the crtc before skipping.  This means that the following
sequence will omit the final CrtcSet call to the driver:

    CRTC c1 connect to output o
    CRTC c2 connect to output o
    CRTC c1 connect to output o

This change adds the check to ensure that each of the calls
are made to the driver.

BUG=chromium:337993
TEST=Test various plug/unplug with DP/HDMI monitors

Change-Id: I5197919ea76204606407dbf72268f9907b71348a
Signed-off-by: David Sodman <dsodman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/184745
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>