platform2: temporarily disable most quipper unit tests
The quipper unit tests are failing when we install the new perf
into the chroot because they rely on the output of perf report
staying constant between versions, but quipper itself doesn't rely
on this.  So, we're temporarily disabling this so that we can upgrade
perf in the chroot.

BUG=chromium:336858
TEST=run unit tests on platform2

Change-Id: I373c5ef5dae20d18e9cd9cac9dc11ffefc170770
Reviewed-on: https://chromium-review.googlesource.com/184785
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ahmad Sharif <asharif@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>