Patch CherryPy to allow binding to an arbitrary available port.
This allows us to pass in a config of server.socket_port == 0 to the
quickstart() method. It requires that the actual port is reported back
up the stack, so that it can be correctly logged on the server's output.
We also care to restore the original bind address upon stopping the
server.

BUG=chromium:333082
TEST=Port allocated and reported with when running devserver with --port=0

Change-Id: I4da5ef4bcef18c1d6d6c23664631716ff613d707
Reviewed-on: https://chromium-review.googlesource.com/184925
Tested-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: Gilad Arnold <garnold@chromium.org>