media-libs/mesa: Add locking to builtin_builder singleton
Consider a multithreaded program with two contexts A and B, and the
following scenario:

1. Context A calls initialize(), which allocates mem_ctx and starts
   building built-ins.
2. Context B calls initialize(), which sees mem_ctx != NULL and assumes
   everything is already set up.  It returns.
3. Context B calls find(), which fails to find the built-in since it
   hasn't been created yet.
4. Context A finally finishes initializing the built-ins.

This will break at step 3.  Adding a lock ensures that subsequent
callers of initialize() will wait until initialization is actually
complete.

Similarly, if any thread calls release while another thread is still
initializing, or calling find(), the mem_ctx/shader would get free'd while
from under it, leading to corruption or use-after-free crashes.

This patch has been sent upstream, but not yet accepted.
It is based upon a patch uploaded for
  https://bugs.freedesktop.org/69200
It fixes a crash in 2014.02.10 piglit test:
  glx-multithread-shader-compile.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:331326
TEST=On link: manually install latest piglit:
    gmerge piglit
TEST=Run previously crashing test:
    /usr/local/piglit/bin/glx-multithread-shader-compile -fbo -auto

Change-Id: Ia237ba258e25546a143b387af9648c951b2943c4
Reviewed-on: https://chromium-review.googlesource.com/185324
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>