Autotest Server Ebuild
Added dependencies we could not add initially.

* Git
* Subversion
* MariaDB
* Apache

Brought in an ebuild for google-web-toolkit and added it to host-hard-depends.

Created a new ebuild to install the autotest files in the rootfs and install
an upstart script to copy it over to stateful.

BUG=chromium:332136,chromium:341749
TEST=Built stumpy_moblab and ensured that the autotest files ended up in the
rootfs on boot and in stateful.

CQ-DEPEND=I799ff93669f21b94d4c216cb756898dcb6d779d7

Change-Id: I15340ddb36d60ba7b4b9847e5e7cd7e3c4e9b447
Reviewed-on: https://chromium-review.googlesource.com/187308
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>