Add the test data of display tests back to public
We are moving the display_Resolution test from private to public.
The test code was done. And this change is to make the test data to
public. The test data is stored in GS and pulled using an ebuild file.

The private ebuild was already removed in:
https://chrome-internal-review.googlesource.com/#/c/152930/3

BUG=chromium:336726
TEST=manaul
$ emerge-link autotest-all
The above command also emerge its RDEPEND autotest-testdata-display.
Checked the test data is installed in:
/build/link/usr/local/build/autotest/server/site_tests/display_Resolution/display_Resolution_test_data

Change-Id: I787884b768699e51c359e1a8d17646549f4e65fa
Reviewed-on: https://chromium-review.googlesource.com/185843
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>