Fix crda and regulatory install issues
The $(get_libdir) functions are useful in some situations but do
not work in the case of the wireless regulatory database.  Despite
mention of a settable "$PUBKEY_DIR" and "$REG_BIN" values in the
crda Makefile, the actual crda code directly references the path
"/usr/lib/crda/regulatory.bin".  As such, the wireless-regdb
package should install there.  Although the pubkey directory can
successfully be moved to /usr/lib64, we might as well share the
same directory for the regulatory.bin and pubkeys.

BUG=chromium:343170
TEST=Reboot DUT and observe messages "cfg80211: World regulatory
domain updated"

Change-Id: I53c6e4702e735f025ab0c0c30b8bec53ea924e0f
Reviewed-on: https://chromium-review.googlesource.com/186141
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>