Add hwid_override USE flag for update_engine
Some boards may not expose HWID in the normal way. This USE flag
turns on functionality in update_engine that causes it to check
/etc/lsb-release for an HWID_OVERRIDE field.

The USE flag causes a compile define (USE_HWID_OVERRIDE) to always
be defined, either to 0 or to 1. This follows the pattern used in
platform2.

BUG=None
TEST=build

Change-Id: I15fdf0149c671757dfdeea5c98d443a55becf3a4
Reviewed-on: https://chromium-review.googlesource.com/186482
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>