Force revbump in chromeos-base to clear out buggy useradd.eclass
There was a bug in useradd.eclass code that caused multiple
copies of entries in /etc/{group,passwd} to get pulled in
if they were present. The eclass was fixed, but binary packages
will have stale copies of the logic. The chromeos-base package
is the only one that uses this eclass, so revving it should fix the issue.

BUG=None
TEST=build_packages

Change-Id: Ibffc61d5322adfb0ceb733add530bc2cc2161031
Reviewed-on: https://chromium-review.googlesource.com/187027
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>