cros-kernel2: Add USE flag for Tegra 5 XHCI controller firmware
The Tegra 5 XHCI controller needs the firmware installed through the
overlay-nyan/sys-kernel/xhci-firmware package. We want to include it in
the built-in firmware filesystem embedded in the kernel image to make it
available in USB boot situations, but since we need to pull it from the
/build/<board>/lib/firmware directory we cannot simply hardcode the
option in chromiumos-tegra124.flavour.config. This patch adds a USE flag
to add it via cros-kernel2, and to make sure the dependent package is
built.

CQ-DEPEND=CL:187513
BUG=None
TEST=USB boot a Nyan after CL:186959 has landed, make sure it works.

Change-Id: I829c8dc4adc7ecd7a2837e60927890c0f243aa88
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/187493
Reviewed-by: Andrew Chew <achew@nvidia.com>
Tested-by: Andrew Chew <achew@nvidia.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>