Make chromeos-base grab a lock on user/group dbs
Since this file modifies group and passwd under the target root, it
needs to grab a lock while it does this in order to be compatible
with the logic in user.eclass. This is temporary, as this ebuild will
soon be updated to rely on the new logic.

BUG=chromium:347938
TEST=build_packages, security_AccountsBaseline
TEST=http://chromegw/p/tryserver.chromiumos/builders/link-release/builds/791

Change-Id: Ia201aae48c7bb7017f00ee8a8677b006317ba828
Reviewed-on: https://chromium-review.googlesource.com/188766
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>