protobuf: Add options to protoc to redirect input/output to files.
When using protoc to encode/decode protobufs, the input/output was from
stdin/stdout only. This patch adds command line flags to specify files to read
the protobuf from and and write the encoded/decoded protobuf to.

This is needed to simplify using this tool in the platform2 build system using
gyp.

BUG=chromium:349252
TEST=(1) emerge-peppy protobuf
     (2) Convert a protobuf between text and binary format on a DUT using the
         new flags.

Change-Id: Ib4d963f096f3232ea39288393bb71000e381e98b
Reviewed-on: https://chromium-review.googlesource.com/189154
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>