Prepare to upgrade FreeType to 2.5.0.1
Freetype 2.5.0.1 or later is required for color font support on Chrome
OS. Noto Color Emoji will be added in a separate CL.

The upstream ebuild can't be cross-compiled as it is. In this CL, we
check in the upstream ebuild but masks it (2.5.0.1). It'll be unmasked
with a local patch in another CL.

BUG=chromium:324184, chromium:245397
TEST=emerge-$BOARD freetype still builds freetype 2.4.12 as before.

Change-Id: I0e5517c8810258a823fdb1893784940b70c628a4
Reviewed-on: https://chromium-review.googlesource.com/189483
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Jungshik Shin <jshin@chromium.org>
Commit-Queue: Jungshik Shin <jshin@chromium.org>