Add a new use 'next_gcc' for early and easy access to next major gcc release.
Toolchain and application developers, who want early access to the
next major gcc release either for performance evaluation or early
validateion, usually use a GITHASH, which changes and is not intuitive
to developers, to access it. We are now to adopt a new 'next_gcc' use
for easy access of next gcc release. So to have the next major gcc
release installed is as easy as -
  USE=next_gcc emerge cross-armv7a-cros-linux-gnueabi/gcc

This has no impact to all builds / nightly tests.

BUG=None
TEST=build 'next_gcc' using this way

Change-Id: I5010c4f42820b5f14ed38b7679e23ca29a795ebc
Reviewed-on: https://chromium-review.googlesource.com/189671
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Tested-by: Han Shen <shenhan@google.com>
Commit-Queue: Han Shen <shenhan@google.com>