Prepare to upgrade Harfbuzz to 0.9.27
Chrome upgraded harfbuzz to 0.9.27 recently. We plan to turn on
use_system_harfbuzz on Chrome OS to avoid a linker warning we treat as
an error. With system_harfbuzz turned on, we have to be in sync with
Chrome tree because harfbuzz in ChromeOS tree will be used for web page
rendering as well as for the UI.

This CL does not actually upgrade harfbuzz to 0.9.27 yet.
Instead, it just takes harfbuzz-0.9.26-r1.ebuild from the upstream
and rename it to harfbuzz-0.9.27-r1.ebuild with KEYWORDS changed to
'~*' to mask on all architectures.

In the follow-up CL, r2 ebuild will be made with one line change to
disable 'gobject' (that leads to a cross-compile error and we don't
need) and mask will be removed and '-introspection' will be added to the
default package.use.

BUG=chromium:353127,chromium:242771,chromium:352902
TEST=NONE (ebuild is masked).

Change-Id: I1777163c889d9d33ff9fb2775d0db5204633780e
Reviewed-on: https://chromium-review.googlesource.com/190893
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jungshik Shin <jshin@chromium.org>
Tested-by: Jungshik Shin <jshin@chromium.org>