Change strongswan-5.0.2-r7.ebuild to a symlink
Follow the pattern used by other ebuilds of this type, where
the non "-r*" ebuild tracks the changes to the ebuild and the
"-r*" file is a symlink.  There is no functional change in
what is actually built, but there is a switch to EAPI=4 to
satisfy pre-submit tests.  Therefore, revbump to -r8.

BUG=chromium:334620
TEST=emerge strongswan, ensure strongswan-5.0.2-r8 is built.

Change-Id: I80c755fd5c9adb156ae09838ef816bc1b2f37daf
Reviewed-on: https://chromium-review.googlesource.com/191096
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>