python: move python-updater to sdk-only
This package makes no sense in a runtime-only world, so move it to the
sdk dependency list.  This avoids pulling in extra deps in the image.

We need to explicitly bump dev-install as it records a list of all the
packages that python needs to bootstrap.

BUG=chromium:311764
TEST=`cbuildbot chromiumos-sdk` works

Change-Id: Ic0930188763ae7ff7b2b6370981a709eb910c845
Reviewed-on: https://chromium-review.googlesource.com/191147
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>