net-wireless/bluez: Enable sixaxis plugin
By enabling the sixaxis plugin, making it possible to connect a PS3
dualshock 3 joystick.

This device is a special case where no actual pairing is required. When
first connected to the USB port, the sixaxis plugin will write the
address of the bluetooth adapter to the joystick firmware and it will
become available in the list of devices of the bluetooth pairing dialog.
When the user presses the PS button on the joystick it will then become
usable as input device.

By disconnecting the USB cable, the joystick will then try to connect
to the adapter which the address was written is on its firmware.

BUG=chromium:359078
TEST=Run bluetoothctl as root and type the following commands:
     power on / agent on / default-agent / scan on
     Plug the joystick to a USB and the device should be available as
     [NEW] Device XX:XX:XX:XX:XX:XX PLAYSTATION(R)3 Controller
     Press the PS button on the joystick, and it can be used via USB.
     Disconnect the cable and the joystick will reconnect wirelessly.
     Quit bluetoothctl and check with btmon command.
CQ-DEPEND=CL:192400

Change-Id: Icf9e4f50768f64c1a8807ad34eff890d30a946a0
Signed-off-by: Eduardo Lima (Etrunko) <eduardo.lima@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/192567
Reviewed-by: Scott James Remnant <keybuk@chromium.org>